+++ src/domain/flag/args.rs
@@ -5,8 +5,8 @@
 
 use anyhow::Result;
 use clap::{Arg, ArgMatches, Command};
-use himalaya_lib::{Flag, Flags};
 use log::{debug, info};
+use pimalaya_email::{Flag, Flags};
 
 use crate::email;